requirements.txt
# Core ML dependencies
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.16.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Unsloth for optimized training
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Additional ML dependencies
scipy>=1.11.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Configuration management
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Experiment tracking (optional but recommended)
wandb>=0.16.0
mlflow>=2.9.0

# Progress bars and CLI
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0

# Evaluation metrics
scikit-learn>=1.3.0
evaluate>=0.4.0

# Memory profiling
psutil>=5.9.0
gputil>=1.4.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0

# Jupyter support (optional)
jupyter>=1.0.0
ipywidgets>=8.1.0
